implemented user login and registration

+++ base/views.py
@@ -1,7 +1,14 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.contrib import messages
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
+
 
 # rooms = [
 #   {'id': 1, 'name': 'Lets learn python!'},
@@ -9,8 +16,58 @@ from .forms import RoomForm
 #   {'id': 3, 'name': 'Devs!'},
 # ]
 
+#! Login
+def loginPage(request):
+  loginpage = True
+  if request.user.is_authenticated:
+    return redirect('home')
 
-# Create your views here.
+  if request.method == 'POST':
+    username = request.POST.get('username').lower()
+    password = request.POST.get('password')
+      
+    try:
+      user = User.objects.get(username=username)
+    except:
+      messages.error(request, "This user does not exist")
+
+    user = authenticate(request, username=username, password=password)
+
+    if user is not None:
+      login(request, user)
+      return redirect('home')
+    else:
+      messages.error(request, 'Password is wrong')
+
+  context = {'loginpage': loginpage}
+
+  return render(request,'base/login_register.html', context)
+
+#! Logout
+def logoutUser(request):
+  logout(request)
+  return redirect('home')
+
+#! Register
+def registerPage(request):
+  form = UserCreationForm()
+
+  if request.method == 'POST':
+    form = UserCreationForm(request.POST)
+    if form.is_valid():
+      user = form.save(commit=False)
+      user.username = user.username.lower()
+      user.save()
+      login(request, user)
+      return redirect('home')
+    else:
+      messages.error(request, 'An error occured during registration')
+
+
+  context = {'form': form}
+  return render(request,'base/login_register.html', context)
+
+#! Home
 def home(request):
   q = request.GET.get('q') if request.GET.get('q') != None else ''
 
@@ -26,6 +83,7 @@ def home(request):
   context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
   return render(request, 'base/home.html', context)
 
+#! Room
 def room(request, pk):
   room = Room.objects.get(id=pk)
   context = {'room': room}
@@ -37,6 +95,8 @@ def room(request, pk):
   #     room = i
   # context = {'room': room}
 
+#! Create Room
+@login_required(login_url="login")
 def createRoom(request):
   form = RoomForm()
   if request.method == 'POST':
@@ -47,9 +107,14 @@ def createRoom(request):
   context = {'form': form}
   return render(request, 'base/room_form.html', context)
 
+#! Update Room
+@login_required(login_url="login")
 def updateRoom(request, pk):
   room = Room.objects.get(id=pk)
   form = RoomForm(instance=room)
+
+  if request.user != room.host:
+    return HttpResponse('You are not allowed here')
 
   if request.method == 'POST':
     form = RoomForm(request.POST, instance=room)
@@ -60,8 +125,13 @@ def updateRoom(request, pk):
   context = {'form': form}
   return render(request, 'base/room_form.html', context)
 
+#! Delete Room
+@login_required(login_url="login")
 def deleteRoom(request, pk):
   room = Room.objects.get(id=pk)
+  
+  if request.user != room.host:
+    return HttpResponse('You are not allowed here')
 
   if request.method == 'POST':
     room.delete()
